requirements-linux.txt
nintendoclients==0.0.3
pycryptodome
pywebview
requests
pygobject
pyclip
pyyaml